pyproject.toml
[project]
name = "unstructured-io"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.8"
dependencies = [
    "meilisearch>=0.40.0",
    "python-dotenv>=1.2.1",
    "temporalio>=1.21.1",
    "unstructured-client>=0.42.8",
]
